ZfsPools with a single ancestor choose ancestor path
When creating a pool that both:

  - Doesn't have a backing partition (mpath setup)
  - Has one backing VDev

The resource_manager chooses the path of the ancestor. This happens here:

https://github.com/whamcloud/integrated-manager-for-lustre/blob/7ff102df7828f4798aa2bda530754960b020cb99/chroma_core/services/plugin_runner/resource_manager.py#L560-L563

This is masking that some devices are pools and presenting a disk id instead.

The real issue here is that the storage API does not implement things in a granular enough capacity to say they are zpools, this is implemented via subclassing in linux.py on the manager.

Instead of looking for `LogicalDriveSlices`, we should expose the device subclasses so we can match on them in the resource_manager.
